Attempt read from ROOM athlete

+++ app/src/main/java/com/company/athleteapiart/MainActivity.kt
@@ -75,7 +75,7 @@ class MainActivity : ComponentActivity() {
                             )
                         ) { entry ->
                             WelcomeScreen(
-                                athleteId = entry.arguments?.getInt("athleteId") ?: -1,
+                                athleteId = entry.arguments?.getLong("athleteId") ?: -1,
                                 accessToken = entry.arguments?.getString("accessToken") ?: "null",
                                 navController = navController
                             )
